delete unused file server 코드를 통합하였고 미사용 파일은 삭제하였습니다.
